Before adding new type of lineage

+++ core/src/main/scala/org/apache/spark/TaskContextImpl.scala
@@ -45,7 +45,7 @@ private[spark] class TaskContextImpl(val stageId: Int,
 
   /** Matteo *************************************************************************************/
   // Used to pipeline records through taps inside the same stage
-  @transient var currentInputId: Int = -1
+  @transient var currentInputId: Long = -1
 
   // Used to pipeline records through taps inside the same stage
   @transient var currentBuffer: ByteBuffer[Long, Int] = null
